+++ components/nav-indicator.ts
