+++ components/splashscreen.ts
