+++ components/welcome.ts
