+++ consts/foo.ts
